Feature Request #413: Refactored added checks for container ephemeral-storage resource request and limit
