+++ score/ephemeralstorage_test.go
@@ -10,27 +10,34 @@ import (
 
 func TestPodContainerStorageEphemeralNoLimit(t *testing.T) {
 	t.Parallel()
-	testExpectedScore(t, "pod-ephemeral-storage-missing-limit.yaml", "Container Ephemeral Storage Requests and Limits", scorecard.GradeCritical)
+	testExpectedScore(t, "pod-ephemeral-storage-missing-limit.yaml", "Container Ephemeral Storage Request and Limit", scorecard.GradeCritical)
 }
 
 func TestPodContainerStorageEphemeralNoRequest(t *testing.T) {
 	t.Parallel()
-	testExpectedScore(t, "pod-ephemeral-storage-missing-request.yaml", "Container Ephemeral Storage Requests and Limits", scorecard.GradeWarning)
+	testExpectedScore(t, "pod-ephemeral-storage-missing-request.yaml", "Container Ephemeral Storage Request and Limit", scorecard.GradeWarning)
 }
 
 func TestPodContainerStorageEphemeralRequestEqualsLimit(t *testing.T) {
 	t.Parallel()
-	testExpectedScore(t, "pod-ephemeral-storage-request-matches-limit.yaml", "Container Ephemeral Storage Requests and Limits", scorecard.GradeAllOK)
+
+	structMap := make(map[string]struct{})
+	structMap["container-ephemeral-storage-request-equals-limit"] = struct{}{}
+
+	testExpectedScoreWithConfig(t, config.Configuration{
+		AllFiles:             []ks.NamedReader{testFile("pod-ephemeral-storage-request-matches-limit.yaml")},
+		EnabledOptionalTests: structMap,
+	}, "Container Ephemeral Storage Request Equals Limit", scorecard.GradeAllOK)
 }
 
 func TestPodContainerStorageEphemeralRequestNoMatchLimit(t *testing.T) {
 	t.Parallel()
 
 	structMap := make(map[string]struct{})
-	structMap["container-ephemeral-storage-request-nomatch-limit"] = struct{}{}
+	structMap["container-ephemeral-storage-request-equals-limit"] = struct{}{}
 
 	testExpectedScoreWithConfig(t, config.Configuration{
 		AllFiles:             []ks.NamedReader{testFile("pod-ephemeral-storage-request-nomatch-limit.yaml")},
 		EnabledOptionalTests: structMap,
-	}, "Container Ephemeral Storage Requests and Limits", scorecard.GradeCritical)
+	}, "Container Ephemeral Storage Request Equals Limit", scorecard.GradeCritical)
 }
